// src/app/core/models/shared/domicilio.model.ts
export class DomicilioModel{
    id: number = null;
    calle: string = null;
    colonia: string = null;
    codigoPostal: string = null;
    municipio: string = null;
    estado: string = null;
    pais: string = null;
    estatus: boolean = true;
    createdAt: Date = null;
    updatedAt: Date = null;
}